+++ packages/admin-ui/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug Fixes
 
 -   `Breadcrumbs`: throw a runtime error when non-last items are missing a `to` prop [#76493](https://github.com/WordPress/gutenberg/pull/76493/)
+-   Fix Page Header not rendering when only `actions` prop is provided. [#76695](https://github.com/WordPress/gutenberg/pull/76695)
 
 ## 1.10.0 (2026-03-18)